+++ assets/javascripts/kitten/components/buttons/button/stories.js
@@ -44,6 +44,7 @@ const modifierOptions = {
   Carbon: 'carbon',
   Oxygen: 'oxygen',
   Checked: 'checked',
+  Copper: 'copper',
 }
 
 const info = {