+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `copper` modifier to `Button`
+
 ## [2.39.0] - 2019-11-12
 
 Features: